+++ .eslintrc.json
@@ -17,7 +17,7 @@
         "@angular-eslint/component-selector": [
           "error",
           {
-            "prefix": ["ng-demo", "ts-svg"],
+            "prefix": ["ng-demo", "rf", "ts-svg"],
             "style": "kebab-case",
             "type": "element"
           }
@@ -25,7 +25,7 @@
         "@angular-eslint/directive-selector": [
           "error",
           {
-            "prefix": ["ngDemo", "tsSvg"],
+            "prefix": ["ngDemo", "rf", "tsSvg"],
             "style": "camelCase",
             "type": "attribute"
           }
