+++ angular.json
@@ -101,43 +101,6 @@
         }
       }
     },
-    "reusable-forms": {
-      "projectType": "library",
-      "root": "projects/reusable-forms",
-      "sourceRoot": "projects/reusable-forms/src",
-      "prefix": "rf",
-      "architect": {
-        "build": {
-          "builder": "@angular-devkit/build-angular:ng-packagr",
-          "options": {
-            "project": "projects/reusable-forms/ng-package.json"
-          },
-          "configurations": {
-            "production": {
-              "tsConfig": "projects/reusable-forms/tsconfig.lib.prod.json"
-            },
-            "development": {
-              "tsConfig": "projects/reusable-forms/tsconfig.lib.json"
-            }
-          },
-          "defaultConfiguration": "production"
-        },
-        "test": {
-          "builder": "@angular-devkit/build-angular:karma",
-          "options": {
-            "main": "projects/reusable-forms/src/test.ts",
-            "tsConfig": "projects/reusable-forms/tsconfig.spec.json",
-            "karmaConfig": "projects/reusable-forms/karma.conf.js"
-          }
-        },
-        "lint": {
-          "builder": "@angular-eslint/builder:lint",
-          "options": {
-            "lintFilePatterns": ["projects/reusable-forms/**/*.ts", "projects/reusable-forms/**/*.html"]
-          }
-        }
-      }
-    },
     "core": {
       "projectType": "library",
       "root": "projects/core",
