+++ package.json
@@ -4,8 +4,7 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve --hmr --host 0.0.0.0 --disable-host-check",
-    "build": "ng build core && npm run build:libs && ng build ng-demo",
-    "build:libs": "ng build reusable-forms",
+    "build": "ng build core && ng build ng-demo",
     "watch": "ng build --watch --configuration development",
     "generate:svgs": "svg-to-ts-files --config ./src/app/projects/treeshakable-svg/.svg-to-tsrc.json",
     "test": "ng test",
