--- projects/reusable-forms/src/lib/components/button/button.component.html
@@ -1,1 +1,0 @@
-<button [type]="type" [attr.disabled]="disabled" (click)="onClick($event)">{{ label }}</button>
