--- projects/reusable-forms/src/public-api.ts
@@ -1,1 +1,0 @@
-export * from './lib/reusable-forms.module';
