+++ src/app/demo/reusable-forms/forms/login-form/login-form.component.html
@@ -1,7 +1,7 @@
 <form [formGroup]="loginForm" (ngSubmit)="onSubmit()">
   <rf-input-email formControlName="email"></rf-input-email>
   <rf-input-password formControlName="password"></rf-input-password>
-  <rf-button type="submit" label="LOG IN" [disabled]="!loginForm.valid"></rf-button>
+  <ng-demo-button type="submit" label="LOG IN" [disabled]="!loginForm.valid"></ng-demo-button>
 
   <div class="form-debug">
     Status : {{ loginForm.status }}<br />
