+++ src/app/demo/reusable-forms/forms/login-form/login-form.component.spec.ts
@@ -1,8 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { ButtonComponent } from '../../components/button/button.component';
-import { InputEmailComponent } from '../../components/input-email/input-email.component';
-import { InputPasswordComponent } from '../../components/input-password/input-password.component';
+import { ButtonComponent } from 'demo/shared';
+import { ReusableFormsModule } from 'reusable-forms';
 
 import { LoginFormComponent } from './login-form.component';
 
@@ -12,8 +11,8 @@ describe('LoginFormComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [LoginFormComponent, InputEmailComponent, InputPasswordComponent, ButtonComponent],
-      imports: [FormsModule, ReactiveFormsModule],
+      declarations: [LoginFormComponent],
+      imports: [FormsModule, ReactiveFormsModule, ReusableFormsModule, ButtonComponent],
     }).compileComponents();
   });
