+++ src/app/demo/reusable-forms/forms/subscribe-form/subscribe-form.component.html
@@ -1,6 +1,6 @@
 <form [formGroup]="subscribeForm" (ngSubmit)="onSubmit()">
   <rf-input-email formControlName="email"></rf-input-email>
-  <rf-button type="submit" label="SUBSCRIBE" [disabled]="!subscribeForm.valid"></rf-button>
+  <ng-demo-button type="submit" label="SUBSCRIBE" [disabled]="!subscribeForm.valid"></ng-demo-button>
 
   <div class="form-debug">
     Status : {{ subscribeForm.status }}<br />
