+++ src/app/demo/reusable-forms/forms/subscribe-form/subscribe-form.component.spec.ts
@@ -1,8 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { ButtonComponent } from '../../components/button/button.component';
-import { InputEmailComponent } from '../../components/input-email/input-email.component';
-import { InputPasswordComponent } from '../../components/input-password/input-password.component';
+import { ButtonComponent } from 'demo/shared';
+import { ReusableFormsModule } from 'reusable-forms';
 
 import { SubscribeFormComponent } from './subscribe-form.component';
 
@@ -12,8 +11,8 @@ describe('SubscribeFormComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [SubscribeFormComponent, InputEmailComponent, InputPasswordComponent, ButtonComponent],
-      imports: [FormsModule, ReactiveFormsModule],
+      declarations: [SubscribeFormComponent],
+      imports: [FormsModule, ReactiveFormsModule, ReusableFormsModule, ButtonComponent],
     }).compileComponents();
   });
