+++ src/app/demo/reusable-forms/reusable-forms.component.html
@@ -11,7 +11,7 @@
 
   <div class="card">
     <div class="card-title">Button standalone</div>
-    <rf-button label="Click me"></rf-button>
+    <ng-demo-button label="Click me"></ng-demo-button>
   </div>
 </div>
