+++ src/app/demo/shared/components/button/button.component.html
@@ -1,0 +1,1 @@
+<button [type]="type" [attr.disabled]="disabled ? '' : null" (click)="onClick($event)">{{ label }}</button>
