+++ src/app/demo/shared/index.ts
@@ -1,0 +1,1 @@
+export * from './components/button/button.component';
