+++ src/app/projects/reusable-forms/public-api.ts
@@ -1,0 +1,2 @@
+export * from './lib/reusable-forms.module';
+export * from './lib/validators/password.validators';
