+++ tsconfig.json
@@ -5,7 +5,8 @@
     "baseUrl": "./",
     "paths": {
       "core": ["dist/core"],
-      "reusable-forms": ["dist/reusable-forms"],
+      "demo/shared": ["src/app/demo/shared"],
+      "reusable-forms": ["src/app/projects/reusable-forms/public-api"],
       "treeshakable-svg": ["src/app/projects/treeshakable-svg/public-api"]
     },
     "outDir": "./dist/out-tsc",
